Full "airesize" functionality for all YCSB-based workloads - Make sure that both the command `airesize <ai name> seed +N` and `airesize <ai name> mongodb +N` results in still functional Cassandra/MongoDB cluster with N additional data nodes. - Applicable to all YCSB AI types: cassandra_ycsb, mongo_ycsb and redis_ycsb - Added a small (AI-specific) utility called `cbcluster`, which will show a snaphsot of the current cluster configuration (for both Cassandra and MongoDB) - IMPORTANT: this commit ensures that the cluster will be properly increased in size and still remains functional. It does not deal with data rebalancing (i.e., it is better to perform the AI resize BEFORE start the load generation on the AI)

+++ scripts/mongo_ycsb/cb_restart_mongo.sh
@@ -27,9 +27,8 @@ do
     ((pos++))
 done
 
-SERVICES[1]="mongodb"
-SERVICES[2]="mongod"
-service_stop_disable ${SERVICES[${LINUX_DISTRO}]}
+service_stop_disable mongod
+service_stop_disable mongodb
 
 # Update Mongo Config
 
@@ -42,19 +41,22 @@ fi
 sudo sed -i "s/bind_ip.*$/bind_ip = ${my_ip_addr}/g" ${MONGODB_CONF_FILE}
 sudo sed -i "s/port.*$/port = 27017/g" ${MONGODB_CONF_FILE}
 
-syslog_netcat "Starting mongod on ${SHORT_HOSTNAME}" 
-
 my_position=$(cat /etc/hosts | grep ${my_ip_addr} | grep mongodb- | tail -1 | awk '{ print $2 }' | sed 's/mongo//g')
 my_dbpath=$(sudo cat /etc/mongodb.conf | grep dbpath | cut -d '=' -f 2)
 
-sudo pkill -9 -f shardsvr
-sudo screen -S MGS -X quit
-sudo screen -d -m -S MGS
-sudo screen -p 0 -S MGS -X stuff "sudo rm /var/lib/mongo/mongod.lock$(printf \\r)"
-sudo screen -p 0 -S MGS -X stuff "mongod --port 27017 --dbpath ${my_dbpath} --shardsvr --replSet cbdrs${my_position} --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log$(printf \\r)"
+sudo ps aux | grep "mongod --port 27017 --dbpath ${my_dbpath} --shardsvr --replSet cbdrs${my_position} --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log" | grep -v grep > /dev/null 2>&1
+if [[ $? -ne 0 ]]
+then
 
-#sudo screen -p 0 -S MGS -X stuff "service mongod restart$(printf \\r)"
+    syslog_netcat "Starting mongod on ${SHORT_HOSTNAME}" 
 
+    sudo pkill -9 -f shardsvr
+    sudo screen -S MGS -X quit
+    sudo screen -d -m -S MGS
+    sudo screen -p 0 -S MGS -X stuff "sudo rm /var/lib/mongo/mongod.lock$(printf \\r)"
+    sudo screen -p 0 -S MGS -X stuff "mongod --port 27017 --dbpath ${my_dbpath} --shardsvr --replSet cbdrs${my_position} --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log$(printf \\r)"
+fi
+    
 sleep 10
 
 sudo screen -S MGSI -X quit
@@ -72,5 +74,8 @@ else
     syslog_netcat "MongoDB server failed to start - NOK"
 fi
 
+check_mongodb_cluster_state ${mongos_ip} cbdrs 0 20
+
 provision_application_stop $START
+
 exit ${STATUS}
